+++ vm/rust/src/juno_state.rs
